pyproject.toml
[tool.poetry]
name = "geo_cache"
version = "0.1.0"
description = "Flask application for geocoding with caching"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
flask = "^2.3.3"
redis = "^5.0.0"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
loguru = "^0.7.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=geo_cache --cov-report=term-missing"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
